feat: update finance components to handle no data state and improve UI
- Removed unused image simbolo.png.
- Enhanced FinancePieChart to display a no data image and message when no data is available.
- Introduced FinancialTransaction component to manage and display financial transaction details with animations.
- Deleted outdated LatestTransactionsCards component and replaced it with a new LatestTransactions component that handles no data state and displays a relevant message.
