+++ package.json
@@ -54,6 +54,7 @@
     "@types/react": "~19.0.10",
     "jest": "^29.2.1",
     "jest-expo": "~53.0.5",
+    "react-native-svg-transformer": "^1.5.1",
     "react-test-renderer": "19.0.0",
     "typescript": "~5.8.3"
   },
